fix(#3607): simulate agent and change test structure
