+++ tests/integration/test_api/test_statistics/data/configuration_template/configuration_manager_statistics_format.yaml
