+++ tests/integration/test_api/test_statistics/data/statistics_template/manager_statistics_format_test_module/wazuh-analysisd_template.json
